backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
pydantic
numpy
scipy
sentence-transformers
faiss-cpu
networkx
pytest
python-dotenv
python-multipart
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pypdf
reportlab
